+++ FOX/recipes/__init__.py
@@ -1,0 +1,13 @@
+"""
+FOX.recipes
+===========
+
+A number of Auto-FOX recipes.
+
+"""
+
+from .param import get_best, overlay_descriptor, plot_descriptor
+
+__all__ = [
+    'get_best', 'overlay_descriptor', 'plot_descriptor'
+]